refactor(test): IIFE 드라이버 헬퍼 writeIifeDriver 공용화 (#3321 /simplify 후속)
#3355 /simplify reuse 에이전트 권고: iife-code-splitting.test.ts 의
writeBrowserDriver + 인라인 __wkr.cjs/__nd.cjs 3개 드라이버가 동일
skeleton(fs/path/here/ev/others→entry) 중복. env 별 스텁 한 줄만 차이.

helpers.ts 에 `writeIifeDriver(dir, entry, allJs, env)` 추출
(env: 'dom'|'worker'|'nondom') — writeOutputs/runNode 등 공용 유틸과
동거. iife 테스트의 writeBrowserDriver 는 'dom' 위임 thin alias 로,
worker/nondom 인라인 블록은 헬퍼 호출로 대체. 드라이버 skeleton 드리프트
방지(PR4 UMD/AMD/RSC 추가 시 재사용).

테스트 전용 — Zig/emit/런타임 무변경. behavior-preserving:
iife-code-splitting 13 pass, helpers.ts 공유 통합(cjs/iife/reexport/
manual-chunks/preserve-modules) 84 pass 0 fail.
